movies/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-dynamic-fixture==4.0.1
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
PyMySQL==1.1.0
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3